Item movers report, halfable orders, floor scale, KOT polish
Four fixes across API + web + mobile:

1. Comprehensive item movers report — new GET /reports/movers (all
   items, category/search/sort/paging + prior-period delta) and
   GET /reports/item/{id} single-item drilldown (trend, by-hour,
   margin). New web Movers page + nav + dashboard "View all" link.

2. Halfable orders — per-item allow_half flag (migration 0044) and
   order_items.qty widened to numeric(6,2). API restricts fractions
   to 0.5 steps on opted-in items; shared formatQty renders ½/3½
   (ASCII for thermal). POS/KDS/history/print updated web + mobile.

3. Floor-menu scale — per-device Compact/Comfortable/Large preset
   (localStorage on web via --pos-scale, MMKV on mobile) plus a
   bigger default; Settings controls on both.

4. KOT polish — minimal cook-docket header (small bold label, no
   doubleSize) so items dominate; slight font bump when printing
   from an Android phone; web/mobile layouts kept consistent.

Tests: 8 new Go integration tests (fractional round-trip, movers,
item drilldown); web tsc/build + mobile tsc/jest green.

+++ apps/api/internal/api/history.go
@@ -43,7 +43,7 @@ type HistoryOrder struct {
 	TaxCents           int64            `json:"tax_cents"`
 	ServiceChargeCents int64            `json:"service_charge_cents"`
 	TotalCents         int64            `json:"total_cents"`
-	ItemCount          int              `json:"item_count"`
+	ItemCount          float64          `json:"item_count"`
 	Items              []OrderItem      `json:"items"`
 	Payments           []HistoryPayment `json:"payments"`
 }
